+++ src/common/components/composite/common/ProfileChip/ProfileChip.tsx
@@ -1,0 +1,60 @@
+import { useTranslation } from 'react-i18next';
+import { gutters } from '../../../../../core/ui/grid/utils';
+import { Box, Button, Skeleton, styled, Tooltip } from '@mui/material';
+import { ProfileView, ProfileViewProps } from './ProfileView';
+import RemoveIcon from '@mui/icons-material/Remove';
+import { FC } from 'react';
+
+export interface ProfileChipProps extends ProfileViewProps {
+  loading?: boolean;
+  removable?: boolean;
+  onRemove?: () => void;
+}
+
+const Root = styled(Box)(({ theme }) => ({
+  borderWidth: 1,
+  borderStyle: 'solid',
+  borderColor: theme.palette.grey.main,
+  borderRadius: theme.shape.borderRadius,
+  paddingLeft: gutters(0.5)(theme),
+  paddingRight: gutters(0.5)(theme),
+}));
+
+const RemoveButton = styled(Button)(({ theme }) => ({
+  minWidth: 'auto',
+  width: gutters(1.5)(theme),
+  height: gutters(1.5)(theme),
+  padding: 0,
+  color: theme.palette.grey.dark,
+  // Put the button on the right of the flex
+  marginLeft: 'auto',
+}));
+
+export const ProfileChip: FC<ProfileChipProps> = ({ loading, removable = false, onRemove, ...props }) => {
+  const { t } = useTranslation();
+
+  return (
+    <Root>
+      {loading && (
+        <Box display="flex" flexDirection="row" alignItems="center" height={gutters(3)} gap={gutters(1)}>
+          <Skeleton variant="circular" sx={{ height: gutters(2), width: gutters(2) }} />
+          <Box flex="1">
+            <Skeleton />
+            <Skeleton />
+          </Box>
+        </Box>
+      )}
+      {!loading && (
+        <ProfileView {...props}>
+          {removable && (
+            <Tooltip title={t('common.remove')} arrow>
+              <RemoveButton onClick={onRemove}>
+                <RemoveIcon />
+              </RemoveButton>
+            </Tooltip>
+          )}
+        </ProfileView>
+      )}
+    </Root>
+  );
+};
